remove TextBuilderMixin
